cOMPILA PERO NO SE :D

+++ vista/src/vista/figura/FiguraView.cpp
@@ -63,7 +63,7 @@ void FiguraView::click(float x, float y) {
 }
 
 void FiguraView::update(Transformacion & tl) {
-	tl.setVector(this->modelo->getX(), this->modelo->getY());
+	tl.setVector((float)this->modelo->getX(), (float)this->modelo->getY());
 	float x,y;
 	tl.getResultadoInverso(x,y);
 	int xi = round(x);
